+++ CHANGES.md
@@ -30,6 +30,12 @@ OpenSSL 3.4
 
 ### Changes between 3.4 and 3.5 [xx XXX xxxx]
 
+ * Changed DTLS handling of invalid records.
+   Records with invalid MAC or invalid length used to generate fatal alert,
+   and are now silently dropped.
+
+   *Omer Kattan*
+
  * Enhanced PKCS#7 inner contents verification.
    In the PKCS7_verify() function, the BIO *indata parameter refers to the
    signed data if the content is detached from p7. Otherwise, indata should be